- Made sign up form operational
