+++ src/routes/signup/+page.server.ts
@@ -1,7 +1,7 @@
 import { superValidate } from 'sveltekit-superforms';
 import { zod } from 'sveltekit-superforms/adapters';
 import type { PageServerLoad } from './$types';
-import { schemaLastStep } from './form/schema';
+import { schemaLastStep } from './schema';
 export const load: PageServerLoad = async () => {
   return {
     form: await superValidate(zod(schemaLastStep)),
